requirements.txt
langchain
pydantic
google-generativeai
pandas
streamlit
